Move more menu to HomeScreen
