+++ phone-app/src/main/java/com/helsinkiwizard/cointoss/ui/MainActivity.kt
@@ -20,14 +20,9 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.systemBarsPadding
-import androidx.compose.foundation.layout.widthIn
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.filled.MoreVert
 import androidx.compose.material3.BottomAppBar
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -43,17 +38,11 @@ import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.NavHostController
@@ -75,7 +64,6 @@ import com.helsinkiwizard.cointoss.utils.AdManager
 import com.helsinkiwizard.core.theme.LocalActivity
 import com.helsinkiwizard.core.theme.ThirtyTwo
 import com.helsinkiwizard.core.theme.TwentyEight
-import com.helsinkiwizard.core.theme.Two
 import com.helsinkiwizard.core.viewmodel.UiState
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
@@ -133,7 +121,7 @@ class MainActivity : ComponentActivity() {
                         LocalActivity provides this@MainActivity,
                         LocalNavController provides navController
                     ) {
-                        CoinToss(navController, isDarkTheme, adsRemoved)
+                        CoinToss(navController, isDarkTheme)
                     }
                 }
 
@@ -160,7 +148,6 @@ class MainActivity : ComponentActivity() {
     private fun CoinToss(
         navController: NavHostController,
         invertColors: Boolean,
-        adsRemoved: Boolean,
     ) {
         val currentDestination = navController.currentBackStackEntryAsState().value?.destination
         val currentRoute = NavRoute.valueOf(currentDestination?.route ?: NavRoute.Home.name)
@@ -203,18 +190,6 @@ class MainActivity : ComponentActivity() {
                         mainGraph()
                     }
                 }
-            }
-
-            AnimatedVisibility(
-                currentRoute in bottomBarItems,
-                modifier = Modifier
-                    .align(Alignment.TopEnd)
-                    .systemBarsPadding()
-            ) {
-                MoreMenu(
-                    navController = navController,
-                    adsRemoved = adsRemoved,
-                )
             }
         }
     }
@@ -317,64 +292,5 @@ class MainActivity : ComponentActivity() {
                 }
             }
         )
-    }
-
-    @Composable
-    private fun MoreMenu(
-        navController: NavHostController,
-        adsRemoved: Boolean
-    ) {
-        var expanded by remember { mutableStateOf(false) }
-
-        IconButton(
-            onClick = { expanded = true }
-        ) {
-            Icon(
-                imageVector = Icons.Default.MoreVert,
-                contentDescription = stringResource(id = R.string.more),
-                tint = MaterialTheme.colorScheme.onSurface
-            )
-        }
-
-        val displayMetrics = LocalConfiguration.current.screenWidthDp.toFloat()
-        val menuWidth = (displayMetrics * CONTEXT_MENU_WIDTH_FRACTION).dp
-
-        DropdownMenu(
-            expanded = expanded,
-            onDismissRequest = { expanded = false },
-            modifier = Modifier.widthIn(min = menuWidth)
-        ) {
-            val menuItems = contextMenuItems.filterNot { adsRemoved && it == NavRoute.RemoveAds }
-            menuItems.forEach { item ->
-                DropdownMenuItem(
-                    text = { Text(stringResource(id = item.titleRes)) },
-                    leadingIcon = {
-                        when {
-                            item.icon != null -> {
-                                Icon(
-                                    imageVector = item.icon,
-                                    contentDescription = null,
-                                    modifier = Modifier.size(TwentyEight)
-                                )
-                            }
-
-                            item.iconRes != null -> {
-                                Icon(
-                                    painter = painterResource(item.iconRes),
-                                    contentDescription = null,
-                                    modifier = Modifier
-                                        .size(TwentyEight)
-                                        .padding(Two)
-                                )
-                            }
-                        }
-                    },
-                    onClick = {
-                        expanded = false
-                        navController.navigate(item.name)
-                    }
-                )
-            }
-        }
     }
 }
